requirements.txt
alchemical==0.5.0
alembic==1.7.5
apifairy==0.9.0
apispec==5.1.1
blinker==1.4
click==8.0.3
Faker==10.0.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
text-unidecode==1.3
webargs==8.0.1
Werkzeug==2.0.2
